+++ packages/antd-ext/components/DatePickerExt/index.ts
@@ -1,3 +1,5 @@
-import Mixin from './Mix';
+import Mixin, { MixinDatePickerOutLine, MixinRangePickerOutLine } from './Mix';
 
 export default Mixin;
+
+export { MixinDatePickerOutLine, MixinRangePickerOutLine };