+++ packages/antd-ext/components/DrawerExt/DrawerExt.tsx
@@ -33,8 +33,6 @@ export default function DrawerExt(props: DrawerExtProps) {
 
   const { prefixCls, mtPrefixCls } = usePrefixCls();
 
-  console.log(okButtonProps, cancelButtonProps);
-
   return (
     <Drawer
       footer={